refact(ImagePuzzleController) : 이름 변경
