+++ Assets/Scripts/Map/Puzzle/ImagePuzzle/ImagePuzzleController.cs.meta
